Add logging callback to ydk
