+++ sdk/cpp/core/src/logger.hpp
@@ -19,9 +19,10 @@
 
 #include <memory>
 #include <string>
-#include "Python.h"
 
 #include "spdlog/spdlog.h"
+#include "logging_callback.hpp"
+
 
 namespace spdlog
 {
@@ -31,128 +32,11 @@ class logger;
 namespace ydk
 {
 
-
-class PyLogger
-{
-public:
-    PyLogger()
-    {
-    }
-
-    ~PyLogger()
-    {
-        if (py_logger != nullptr)
-        {
-           Py_DECREF(py_logger);
-        }
-    }
-
-    void set_logger()
-    {
-        PyObject* py_logging  = PyImport_ImportModule("logging");
-        PyObject* ydk_arg = Py_BuildValue("s", "ydk");
-        const char* method = "getLogger";
-        const char* fmt = "O";
-        py_logger = PyObject_CallMethod(py_logging, (char *)method, (char *)fmt, ydk_arg);
-
-        // Add null handler suppress
-        // `No handlers could be found for logger "ydk"` error message.
-        // This check is added after Python 2.7.4rc1.
-        #if 0x02070000 <= PY_VERSION_HEX
-        const char* null_method = "NullHandler";
-        PyObject* null_handler = PyObject_CallMethod(py_logging, (char *)null_method, NULL);
-        const char* append_method = "addHandler";
-        PyObject_CallMethod(py_logger, (char *)append_method, (char *)fmt, null_handler);
-        Py_DECREF(null_handler);
-        #endif
-
-        Py_DECREF(ydk_arg);
-        Py_DECREF(py_logging);
-    }
-
-    template <typename... Args> void py_fmt_log(const std::string& name, const char* fmt, spdlog::level::level_enum lvl, const char* py_lvl, const Args&... args)
-    {
-        spdlog::details::log_msg log_msg(&name, lvl);
-        log_msg.raw.write(fmt, args...);
-        py_log(log_msg.raw.c_str(), py_lvl);
-    }
-
-    template <typename... Args> void trace(const std::string& name, const char* fmt, const Args&... args) {
-        const char* py_lvl = "debug";
-        py_fmt_log(name, fmt, spdlog::level::trace, py_lvl, args...);
-    }
-    template <typename... Args> void debug(const std::string& name, const char* fmt, const Args&... args) {
-        const char* py_lvl = "debug";
-        py_fmt_log(name, fmt, spdlog::level::trace, py_lvl, args...);
-    }
-    template <typename... Args> void info(const std::string& name, const char* fmt, const Args&... args) {
-        const char* py_lvl = "info";
-        py_fmt_log(name, fmt, spdlog::level::trace, py_lvl, args...);
-    }
-    template <typename... Args> void warn(const std::string& name, const char* fmt, const Args&... args) {
-        const char* py_lvl = "warn";
-        py_fmt_log(name, fmt, spdlog::level::trace, py_lvl, args...);
-    }
-    template <typename... Args> void error(const std::string& name, const char* fmt, const Args&... args) {
-        const char* py_lvl = "error";
-        py_fmt_log(name, fmt, spdlog::level::trace, py_lvl, args...);
-    }
-    template <typename... Args> void critical(const std::string& name, const char* fmt, const Args&... args) {
-        const char* py_lvl = "critical";
-        py_fmt_log(name, fmt, spdlog::level::trace, py_lvl, args...);
-    }
-
-    template <typename T> void py_log(const T& msg, const char* py_lvl)
-    {
-        PyObject* py_msg_arg = Py_BuildValue("s", msg);
-        const char* fmt = "O";
-        if (Py_IsInitialized())
-        {
-            if (py_logger == nullptr)
-            {
-                set_logger();
-            }
-            PyObject_CallMethod(py_logger, (char *)py_lvl, (char *)fmt, py_msg_arg);
-        }
-        Py_DECREF(py_msg_arg);
-    }
-
-    template <typename T> void trace(const T& msg) {
-        const char* py_lvl = "debug";
-        py_log(msg, py_lvl);
-    }
-    template <typename T> void debug(const T& msg) {
-        const char* py_lvl = "debug";
-        py_log(msg, py_lvl);
-    }
-    template <typename T> void info(const T& msg) {
-        const char* py_lvl = "info";
-        py_log(msg, py_lvl);
-    }
-    template <typename T> void warn(const T& msg) {
-        const char* py_lvl = "warn";
-        py_log(msg, py_lvl);
-    }
-    template <typename T> void error(const T& msg) {
-        const char* py_lvl = "error";
-        py_log(msg, py_lvl);
-    }
-    template <typename T> void critical(const T& msg) {
-        const char* py_lvl = "critical";
-        py_log(msg, py_lvl);
-    }
-
-
-private:
-    PyObject* py_logger = nullptr;
-};
-
 class Logger
 {
     public:
         Logger()
-            : internal_logger{ spdlog::get("ydk") },
-              py_logger{ std::make_shared<PyLogger>() }
+            : internal_logger{ spdlog::get("ydk") }
         {
         }
 
@@ -164,20 +48,28 @@ class Logger
         template <typename... Args> \
         void loglevel(const char* fmt, const Args&... args) \
         { \
+            logging_callback func = get_logging_callback();\
+            if(func != nullptr) \
+            { \
+                (*func)(1, fmt); //TODO the level needs to be set correctly and formatted string needs to be passed \
+                return;\
+            } \
             if(!lazy_check()) { return; } \
             internal_logger->loglevel<Args...>(fmt, args...); \
-            if (Py_IsInitialized()) \
-            py_logger->loglevel<Args...>(internal_logger->name(), fmt, args...); \
         }
 
         #define YDKLOGLEVELNOARGS(loglevel) \
         template <typename T> \
         void loglevel(const T& msg) \
         { \
+            logging_callback func = get_logging_callback();\
+            if(func != nullptr) \
+            { \
+                (*func)(1, msg.c_str()); //TODO the level needs to be set correctly and formatted string needs to be passed\
+                return;\
+            } \
             if(!lazy_check()) { return; } \
             internal_logger->loglevel<T>(msg); \
-            if (Py_IsInitialized()) \
-            py_logger->loglevel<T>(msg); \
         }
 
         YDKLOGLEVELARGS(trace)
@@ -218,7 +110,6 @@ class Logger
 
     private:
         std::shared_ptr<spdlog::logger> internal_logger;
-        std::shared_ptr<PyLogger> py_logger;
 };
 
 static Logger logger{};
