+++ sdk/cpp/core/src/logging_callback.cpp
@@ -1,0 +1,32 @@
+/*  ----------------------------------------------------------------
+ Copyright 2017 Cisco Systems
+
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+------------------------------------------------------------------*/
+
+#include "logging_callback.hpp"
+
+namespace ydk
+{
+static logging_callback ydk_logging_function = nullptr;
+
+void set_logging_callback(logging_callback func)
+{
+    ydk_logging_function = func;
+}
+
+logging_callback get_logging_callback()
+{
+    return ydk_logging_function;
+}
+}
